Create README.md
We have cleaned the extracted data by removing the outliers and skewness of the data and filtered the data by using categorical and numerical variables, encoding the categorical data, transforming the numerical data, and scaling the overall data using the minmax scaler and training it using different classification and regression models. Finalizing each model based on the hypertuning of the parameters. Then save the models using the Pickle module . Then we are creating the streamlit for the prediction of the copper price.